refactor(sdk): extract security types into advanced_options module
Mirror the core crate's advanced_options extraction in both SDKs:
- Python: PySecurityOptions, PyAdvancedBoxOptions → advanced_options.rs
- Node: JsSecurityOptions, coerce helpers, tests → advanced_options.rs
